kf-slam converted into app class
