+++ apps/kf-slam/CMakeLists.txt
@@ -25,6 +25,6 @@ target_link_libraries(${PROJECT_NAME} ${MRPT_LINKER_LIBS})
 # Dependencies on MRPT libraries:
 #  Just mention the top-level dependency, the rest will be detected automatically,
 #  and all the needed #include<> dirs added (see the script DeclareAppDependencies.cmake for further details)
-DeclareAppDependencies(${PROJECT_NAME} mrpt-slam mrpt-gui mrpt-bayes)
+DeclareAppDependencies(${PROJECT_NAME} mrpt-apps)
 
 DeclareAppForInstall(${PROJECT_NAME})
