+++ libs/apps/CMakeLists.txt
@@ -5,8 +5,9 @@ define_mrpt_lib(
 	# Lib name
 	apps
 	# Dependencies
+	mrpt-bayes
+	mrpt-gui
 	mrpt-obs
 	mrpt-slam
-	mrpt-gui
 	mrpt-tclap
 	)
